Updated config.yml - remove caching
This caching of "dependencies" is somehow caching the entire project,
which breaks imports of new functions, which aren't in the cache.  I
don't understand any of this and it seems wrong to cache things that can
change whether tests fail or pass, so I will just leave it disabled and
hopefully it will work now?