Fix _random_tiebreak for n > 1
For example, `_random_tiebreak([0, 1], 3)` produces
ValueError: Sample larger than population or is negative

Removing the length check breaks the test for runoff voting with a
unanimous electorate (when there is only 1 high scorer, but the
tiebreaker is asked for 2). Maybe there's a more elegant way to deal
with that case, but this works for now.