Separate _keep and _elim tiebreaker functions
This seems like the best way to combine them.  Adding a keep=True
parameter would need to be added to all of them, even though it only
affects two systems.  Inverting the indices and uninverting them to
keep the order backwards clutters up the code when working with lists,
and probably has extra computation costs.

With this method, dict will just be different for each method, which
might be good anyway in case there are method-specific tiebreakers.