Add Approval, Borda, Coombs methods
"…if our goal were simply to design a voting system that has a
moderating effect, there are many additional systems not in wide use
that exhibit this property (any Condorcet method, for instance, or the
Coombs rule."

"…there are voting systems that always select the most moderate
candidate with symmetric 1-Euclidean voters. This is true for any system
that satisfies the Condorcet criterion, selecting the Condorcet winner
whenever one exists (this holds for the Copeland, Black, and Dodgson
methods); it is also true for some other voting systems that do not in
general satisfy the Condorcet criterion, like the Coombs rule."

"There has been some theoretical work on moderating effects of
score-based voting systems (like Borda count and approval voting) with
strategic voters and candidates. However, it is an open question
(with some computational evidence to support it) whether other
voting systems like Borda count exert a moderating effect in the setting
we study, with fixed voter and candidate distributions."

(This file already contains Black and STAR, another score-based system.)